Move add_promo in views.py
