+++ back/upont/urls.py
@@ -13,22 +13,21 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-import django_cas_ng.views
 from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib import admin
 from django.contrib.auth import views as auth_views
 from django.urls import include, path
-from upont.add_promo import add
 
 from . import views
+import django_cas_ng.views
 
 urlpatterns = [
     path("social/", include("social.urls")),
     path("news/", include("news.urls")),
     path("admin/", admin.site.urls),
     path("tellme/", include("tellme.urls"), name="tellme"),
-    path("add_promo/", add, name="add_promo"),
+    path("add_promo/", views.add, name="add_promo"),
     path(
         "login/",
         auth_views.LoginView.as_view(redirect_authenticated_user=True),
